requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
opencv-python==4.8.0.76
Pillow==10.0.0
requests==2.31.0
twilio==8.5.0
sendgrid==6.10.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
schedule==1.2.0
rasterio==1.3.8
geopandas==0.13.2
folium==0.14.0
plotly==5.15.0
dash==2.13.0
celery==5.3.1
redis==4.6.0